refactoring: separate methods from bank_transaction model to concern
